backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.2
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0
aiosqlite==0.20.0
authlib==1.3.1
itsdangerous==2.2.0
python-dotenv==1.2.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
uuid==1.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
slowapi==0.1.9
structlog==24.4.0
